#1521 - Merged the two versions of BirthDateDtos
