+++ sormas-ui/src/main/java/de/symeda/sormas/ui/contact/components/linelisting/LineListingLayout.java
@@ -23,6 +23,7 @@ import com.vaadin.ui.themes.ValoTheme;
 
 import de.symeda.sormas.api.Disease;
 import de.symeda.sormas.api.FacadeProvider;
+import de.symeda.sormas.api.caze.BirthDateDto;
 import de.symeda.sormas.api.caze.CaseDataDto;
 import de.symeda.sormas.api.caze.CaseReferenceDto;
 import de.symeda.sormas.api.contact.ContactDto;
@@ -41,7 +42,6 @@ import de.symeda.sormas.ui.UserProvider;
 import de.symeda.sormas.ui.utils.ButtonHelper;
 import de.symeda.sormas.ui.utils.CssStyles;
 import de.symeda.sormas.ui.utils.FieldHelper;
-import de.symeda.sormas.ui.utils.components.birthdate.BirthDateDto;
 import de.symeda.sormas.ui.utils.components.birthdate.BirthDateField;
 
 public class LineListingLayout extends VerticalLayout {
